@@ -2,7 +2,12 @@ import React from 'react';
 import './App.css';
 
 function App() {
-  return <div className="App"></div>;
+	return (
+		<div className='App'>
+			<h1>Hello World</h1>
+			<h2>Time to start coding ⚛</h2>
+		</div>
+	);
 }
 
 export default App;
